Revert "Fixed clippy lints"
This reverts commit 62584b77fe72f804995eaf5343a9d82ac4cac1cc.

+++ rust/template/cmd_parser/lib.rs
@@ -28,7 +28,7 @@ where
     let mut buf: Vec<u8> = Vec::new();
 
     let istty = unsafe {
-        libc::isatty(/*libc::STDIN_FILENO*/ 0)
+        libc::isatty(/*libc::STDIN_FILENO*/ 0 as i32)
     } != 0;
     let mut input = if istty {
         let mut rl = Editor::<()>::new();
